# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetCare Pro API - Logiciel vétérinaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: Liste des patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: espece
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientWithOwner"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Créer un patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Patient créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Détails d'un patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientWithOwner"
        "404":
          description: Non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Mettre à jour un patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "200":
          description: Patient mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Supprimer un patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Supprimé
  /patients/{id}/consultations:
    get:
      operationId: listPatientConsultations
      tags: [patients]
      summary: Historique des consultations d'un patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultations du patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Owner:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        email:
          type: ["string", "null"]
        telephone:
          type: string
        adresse:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - prenom
        - telephone
        - createdAt
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        espece:
          type: string
        race:
          type: ["string", "null"]
        sexe:
          type: string
        dateNaissance:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        couleur:
          type: ["string", "null"]
        sterilise:
          type: boolean
        ownerId:
          type: integer
        antecedents:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - espece
        - sexe
        - sterilise
        - ownerId
        - createdAt
    PatientWithOwner:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            owner:
              $ref: "#/components/schemas/Owner"
    CreatePatientBody:
      type: object
      properties:
        nom:
          type: string
        espece:
          type: string
        race:
          type: ["string", "null"]
        sexe:
          type: string
        dateNaissance:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        couleur:
          type: ["string", "null"]
        sterilise:
          type: boolean
        ownerId:
          type: integer
        antecedents:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
      required:
        - nom
        - espece
        - sexe
        - sterilise
        - ownerId
    Consultation:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        veterinaire:
          type: string
        date:
          type: string
        statut:
          type: string
        motif:
          type: ["string", "null"]
        anamnese:
          type: ["string", "null"]
        examenClinique:
          type: ["string", "null"]
        examensComplementaires:
          type: ["string", "null"]
        diagnostic:
          type: ["string", "null"]
        diagnosticIA:
          type: ["string", "null"]
        ordonnance:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - veterinaire
        - date
        - statut
        - createdAt
...